+++ src/groupeddataframe/groupeddataframe.jl
@@ -531,8 +531,16 @@ end
 Base.axes(gd::GroupedDataFrame, i::Integer) = Base.OneTo(size(gd, i))
 
 Base.first(gd::GroupedDataFrame) = gd[1]
-Base.last(gd::GroupedDataFrame) = gd[end]
+function Base.first(gd::GroupedDataFrame, n::Integer)
+    n < 0 && throw(ArgumentError("Number of elements must be nonnegative"))
+    return gd[1:min(n, end)]
+end
 
+Base.last(gd::GroupedDataFrame) = gd[end]
+function Base.last(gd::GroupedDataFrame, n::Integer)
+    n < 0 && throw(ArgumentError("Number of elements must be nonnegative"))
+    return gd[max(1, end - n + 1):end]
+end
 # These have to be defined for some to_indices() logic to work, as long
 # as GroupedDataFrame is not <: AbstractArray
 Base.IndexStyle(::Type{<:GroupedDataFrame}) = IndexLinear()