+++ NEWS.md
@@ -15,6 +15,8 @@
 * Add support for `reverse!`, `permute!`, `invpermute!`, `shuffle`,
   and `shuffle!` functions. Improve functionality of `reverse`.
   ([#3010](https://github.com/JuliaData/DataFrames.jl/pull/3010)).
+* `first` and `last` for `GroupedDataFrame` now support passing number of elements to get
+  ([#3006](https://github.com/JuliaData/DataFrames.jl/issues/3006))
 
 # DataFrames.jl v1.3.2 Patch Release Notes